cleaned up server.js, removed manifest.json from index.html, updated Contact page with email, phone and location, styled contact p tag, added mailer.route with updated code to handle the contact form
