+++ Server/server.js
@@ -1,90 +1,13 @@
-require('./config/db')();
 const express = require('express');
-const apiRoutes = require('./routes/api.routes');
 const app = express();
+const apiRoutes = require('./routes/api.routes');
+const mailerRoute = require('./routes/mailer.route');
 const passport = require("passport");
 const logger = require("morgan");
 const User = require("./models/User");
 const LocalStrategy = require("passport-local").Strategy;
-require('dotenv').config();
-
-var router = express.Router();
-var nodemailer = require('nodemailer');
-var cors = require('cors');
-app.use(express.json());
-
-const transport = {
-  host: 'smtp.zoho.com', // Don’t forget to replace with the SMTP host of your provider
-  port: 465,
-  secure: true,
-  auth: {
-  user: process.env.CREDENTIAL_USER,
-  pass: process.env.CREDENTIAL_PASS
-}
-}
-
-const transporter = nodemailer.createTransport(transport)
-
-transporter.verify((error, success) => {
-if (error) {
-  console.log(error);
-} else {
-  console.log('Server is ready to take messages');
-}
-})
-
-router.post('/send', (req, res, next) => {
-  var name = req.body.name
-  console.log(name);
-  var email = req.body.email
-  var message = req.body.message
-  var content = `name: ${name} \n email: ${email} \n message: ${message} `
-
-  var mail = {
-    from: 'surev0te@zohomail.com',
-    to: 'surev0te@zohomail.com',  // Change to email address that you want to receive messages on
-    subject: 'New Message from Contact Form',
-    text: content
-  }
-
-  var mail2 = {
-    from: 'surev0te@zohomail.com',
-    to: email,
-    subject: 'A message from the team',
-    text: 'Here is a copy of your message' +  message  + 'Thanks for your feedback. We will get your issue resolved as soon as possible. - The SureVote Team'
-  }
-  
-  transporter.sendMail(mail, (err, data) => {
-    if (err) {
-      res.json({
-        status: 'fail'
-      })
-      console.log("mail", err);
-    } else {
-      res.json({
-       status: 'success'
-      })
-    }
-  })
-
-  transporter.sendMail(mail2, (err, data) => {
-    if (err) {
-      res.json({
-        status: 'fail'
-      })
-      console.log("mail2", err);
-    } else {
-      res.json({
-       status: 'success'
-      })
-    }
-  })
-});
-
-app.use(cors())
-app.use('/', router)
-
-
+const cors = require('cors');
+require('./config/db')();
 
 app.use(logger("dev"));
 
@@ -105,9 +28,11 @@ passport.use(new LocalStrategy(User.authenticate()));
 
 // parsing middleware
 app.use(express.urlencoded({ extended: true }));
-// app.use(express.json());
-app.use(apiRoutes);
+app.use(express.json());
+app.use(cors())
 
+app.use(apiRoutes);
+app.use(mailerRoute);
 // fetch data from back end to profile path
 // app.get('/profile', (req, res) => {
 //   console.log(__dirname = '/controllers')
